vfs-visa-bot/requirements.txt
selenium==4.15.2
undetected-chromedriver==3.5.4
python-dotenv==1.0.0
requests==2.31.0
telebot==0.0.5
pyTelegramBotAPI==4.14.0
beautifulsoup4==4.12.2
lxml==4.9.3
Pillow==10.1.0
webdriver-manager==4.0.1

